requirements.txt
pycore_utils
flask-socketio
pillow
cryptography
pycryptodome
requests
boto3